requirements.txt
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
mysql-connector==2.2.9
pycodestyle==2.6.0
requests==2.23.0
urllib3==1.25.9
